requirements.txt
# QUANTUM ALPHA - Python Dependencies
# 세계 최고 수준 암호화폐 선물 자동매매 시스템

# ============================================================================
# Core Framework
# ============================================================================
python>=3.11

# ============================================================================
# Deep Learning & AI
# ============================================================================
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.1.0
onnx>=1.15.0
onnxruntime>=1.16.0

# Model utilities
transformers>=4.35.0
timm>=0.9.12

# ============================================================================
# Data Processing & Analysis
# ============================================================================
numpy>=1.24.0
pandas>=2.1.0
polars>=0.19.0
pyarrow>=14.0.0

# Time series
statsmodels>=0.14.0
scipy>=1.11.0

# Signal processing
pywavelets>=1.4.1

# ============================================================================
# Database & Caching
# ============================================================================
# PostgreSQL / TimescaleDB
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# Redis
redis>=5.0.0
hiredis>=2.2.3

# ============================================================================
# Exchange APIs
# ============================================================================
ccxt>=4.1.0
python-bybit>=3.0.0
python-binance>=1.0.19

# HTTP client
httpx>=0.25.0
aiohttp>=3.9.0

# WebSocket
websockets>=12.0

# ============================================================================
# Feature Engineering
# ============================================================================
# Technical indicators
ta-lib>=0.4.28  # May require manual installation
pandas-ta>=0.3.14b0

# Fractional differencing
numba>=0.58.0

# ============================================================================
# Machine Learning
# ============================================================================
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
catboost>=1.2.0

# Clustering
hdbscan>=0.8.33

# ============================================================================
# Backtesting
# ============================================================================
backtrader>=1.9.78
vectorbt>=0.26.0

# ============================================================================
# Monitoring & Logging
# ============================================================================
loguru>=0.7.2
prometheus-client>=0.19.0
grafana-api>=1.0.3

# ============================================================================
# Configuration
# ============================================================================
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.5.0
dynaconf>=3.2.4

# ============================================================================
# Utilities
# ============================================================================
# Date/Time
python-dateutil>=2.8.2
pytz>=2023.3

# UUID
uuid>=1.30

# Cryptography
cryptography>=41.0.7
pycryptodome>=3.19.0

# ============================================================================
# Async & Concurrency
# ============================================================================
asyncio>=3.4.3
aiodns>=3.1.1
uvloop>=0.19.0  # For faster asyncio

# ============================================================================
# Alerts & Notifications
# ============================================================================
python-telegram-bot>=20.7
discord-webhook>=1.3.1

# Email
aiosmtplib>=3.0.1

# ============================================================================
# Testing
# ============================================================================
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0
hypothesis>=6.92.1

# ============================================================================
# Code Quality
# ============================================================================
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.7.1
pylint>=3.0.3

# ============================================================================
# Jupyter & Visualization (Development)
# ============================================================================
jupyter>=1.0.0
ipython>=8.18.0
ipywidgets>=8.1.1

matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
mplfinance>=0.12.10

# ============================================================================
# Performance Profiling
# ============================================================================
line-profiler>=4.1.1
memory-profiler>=0.61.0
py-spy>=0.3.14

# ============================================================================
# Documentation
# ============================================================================
sphinx>=7.2.6
sphinx-rtd-theme>=2.0.0
mkdocs>=1.5.3

# ============================================================================
# Additional Libraries
# ============================================================================
# Progress bars
tqdm>=4.66.1
rich>=13.7.0

# CLI
click>=8.1.7
typer>=0.9.0

# JSON
orjson>=3.9.10  # Faster JSON

# ============================================================================
# Rust Integration
# ============================================================================
# For Python-Rust bridge if needed
maturin>=1.4.0

# ============================================================================
# Notes
# ============================================================================
# Installation tips:
#
# 1. Create virtual environment:
#    python -m venv venv
#    source venv/bin/activate  # On Windows: venv\Scripts\activate
#
# 2. Install PyTorch with CUDA support (if available):
#    pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#
# 3. Install TA-Lib (requires separate installation):
#    - Ubuntu: sudo apt-get install ta-lib
#    - macOS: brew install ta-lib
#    - Windows: Download from https://www.lfd.uci.edu/~gohlke/pythonlibs/#ta-lib
#
# 4. Install all requirements:
#    pip install -r requirements.txt
#
# 5. For development with Jupyter:
#    pip install -r requirements.txt[dev]
